Revert "[USERS-1] [Fixtures] [Part 2] Add script to load users fixture (#12)"
This reverts commit c920610313e479e720bda9fc95c6036f354027d8.
