+++ .gitignore
@@ -23,7 +23,6 @@ downloads/
 eggs/
 .eggs/
 lib/
-!scripts/lib
 lib64/
 parts/
 sdist/
