--- scripts/__init__.py
@@ -1,4 +1,0 @@
-import os
-import sys
-
-sys.path.append(os.path.dirname(os.path.realpath(__file__)))
